refactor: 🚚 rename controller class file
